+++ backend/django_project/settings.py
@@ -167,3 +167,10 @@ DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 REDIS_HOST = env.str("REDIS_HOST", "redis")
 REDIS_PORT = env.int("REDIS_PORT", 6379)
 REDIS_DB = env.int("REDIS_DB", 0)
+
+# Celery
+CELERY_BROKER_URL = "redis://redis:6379"
+
+# Telegram
+# These env vars must be set for worker service in docker-compose.yml
+TELEGRAM_BOT_TOKEN = env.str("TELEGRAM_BOT_TOKEN", None)